Code cleanup. Nearly there with regex escaping
